requirements.txt
aiohttp==3.10.5
APScheduler==3.10.4
dateparser==1.2.0
feedparser==6.0.11
Flask==3.0.3
pandas==2.2.2
pymongo==4.8.0
python-dotenv==1.0.1
gunicorn
